+++ packages/snap/src/utils/executeOnChainActions.ts
@@ -2,7 +2,7 @@ import type { TronWeb } from 'tronweb';
 
 import type { SnapClient } from '../clients/snap/SnapClient';
 import type { TronWebFactory } from '../clients/tronweb/TronWebFactory';
-import type { Network } from '../constants';
+import { type Network, ACCOUNT_SYNC_DELAY } from '../constants';
 import type { TronKeyringAccount } from '../entities/keyring-account';
 import { BackgroundEventMethod } from '../handlers/cronjob';
 import type { AccountsService } from '../services/accounts/AccountsService';
@@ -53,8 +53,7 @@ export async function executeOnChainActions({
   }
 
   await snapClient.scheduleBackgroundEvent({
-    method: BackgroundEventMethod.SynchronizeAccount,
-    params: { accountId: account.id },
-    duration: 'PT5S',
+    method: BackgroundEventMethod.SynchronizeSelectedAccounts,
+    duration: ACCOUNT_SYNC_DELAY,
   });
 }